# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "minotari",
    "description": "",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/accounts/{name}/completed_transactions": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves completed transactions for a specified account with pagination.",
        "description": "Returns a paginated list of completed transactions for the account, including\ntheir status, mined block information, and confirmation details. Transactions\nare ordered by creation time (most recent first).\n\n# Path Parameters\n\n- `name`: The unique account name to query\n\n# Query Parameters\n\n- `limit`: Maximum number of transactions to return (default: 50, max: 1000)\n- `offset`: Number of transactions to skip for pagination (default: 0)\n\n# Response\n\nReturns a list of [`CompletedTransactionResponse`] objects, each containing:\n- Transaction ID and status\n- Kernel excess (hex encoded)\n- Mining and confirmation details\n- Creation and update timestamps\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Request\n\n```bash\n# Get first 50 completed transactions (default)\ncurl -X GET http://localhost:8080/accounts/default/completed_transactions \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n\n# Get 100 transactions starting from offset 50\ncurl -X GET \"http://localhost:8080/accounts/default/completed_transactions?limit=100&offset=50\" \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n```\n\n# Example Response\n\n```json\n[\n  {\n    \"id\": \"550e8400-e29b-41d4-a716-446655440000\",\n    \"pending_tx_id\": \"661e8400-e29b-41d4-a716-446655440001\",\n    \"account_id\": 1,\n    \"status\": \"mined_confirmed\",\n    \"last_rejected_reason\": null,\n    \"kernel_excess_hex\": \"abc123...\",\n    \"sent_payref\": \"payref-123\",\n    \"sent_output_hash\": \"def456...\",\n    \"mined_height\": 12345,\n    \"mined_block_hash_hex\": \"789abc...\",\n    \"confirmation_height\": 12350,\n    \"broadcast_attempts\": 1,\n    \"created_at\": \"2024-01-15T10:30:00Z\",\n    \"updated_at\": \"2024-01-15T10:35:00Z\"\n  }\n]\n```",
        "operationId": "api_get_completed_transactions",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve completed transactions for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of transactions to return (default: 50, max: 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of transactions to skip for pagination (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completed transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompletedTransactionResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{name}/completed_transactions/by_payref/{payref}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves a completed transaction by its payment reference.",
        "description": "Returns a completed transaction that matches the specified payment reference.\nThe payment reference is typically assigned when a transaction is confirmed\non the blockchain.\n\n# Path Parameters\n\n- `name`: The unique account name to query\n- `payref`: The payment reference to search for\n\n# Response\n\nReturns a [`CompletedTransactionResponse`] object if found, containing:\n- Transaction ID and status\n- Kernel excess (hex encoded)\n- Mining and confirmation details\n- Creation and update timestamps\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::NotFound`]: No transaction found with the given payment reference\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Request\n\n```bash\ncurl -X GET http://localhost:8080/accounts/default/completed_transactions/by_payref/my-payment-ref-123 \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n```",
        "operationId": "api_get_completed_transaction_by_payref",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve transaction from",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "payref",
            "in": "path",
            "description": "Payment reference to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completed transaction retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletedTransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account or transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts/{name}/displayed_transactions": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves displayed transactions for a specified account with pagination.",
        "description": "Returns a paginated list of user-friendly transactions for the account,\nincluding incoming and outgoing transactions with their status, amounts,\nand blockchain information. Transactions are ordered by block height\n(most recent first).\n\n# Path Parameters\n\n- `name`: The unique account name to query\n\n# Query Parameters\n\n- `limit`: Maximum number of transactions to return (default: 50, max: 1000)\n- `offset`: Number of transactions to skip for pagination (default: 0)\n\n# Response\n\nReturns a list of [`DisplayedTransaction`] objects, each containing:\n- Transaction ID, direction (incoming/outgoing), and source\n- Status (pending, unconfirmed, confirmed, cancelled, etc.)\n- Amount and formatted display amount\n- Counterparty information (if available)\n- Blockchain details (block height, timestamp, confirmations)\n- Fee information (for outgoing transactions)\n- Detailed input/output information\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Request\n\n```bash\n# Get first 50 displayed transactions (default)\ncurl -X GET http://localhost:8080/accounts/default/displayed_transactions \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n\n# Get 100 transactions starting from offset 50\ncurl -X GET \"http://localhost:8080/accounts/default/displayed_transactions?limit=100&offset=50\" \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n```",
        "operationId": "api_get_displayed_transactions",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve displayed transactions for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of transactions to return (default: 50, max: 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of transactions to skip for pagination (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Displayed transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayedTransaction"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{name}/displayed_transactions/by_payref/{payref}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves displayed transactions by payment reference.",
        "description": "Returns displayed transactions that contain the specified payment reference.\nThis searches within the payment references stored in each transaction.\n\n# Path Parameters\n\n- `name`: The unique account name to query\n- `payref`: The payment reference to search for\n\n# Response\n\nReturns a list of [`DisplayedTransaction`] objects that contain the payment reference.\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Request\n\n```bash\ncurl -X GET http://localhost:8080/accounts/default/displayed_transactions/by_payref/my-payment-ref-123 \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n```",
        "operationId": "api_get_displayed_transactions_by_payref",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve transactions from",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "payref",
            "in": "path",
            "description": "Payment reference to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Displayed transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayedTransaction"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompletedTransactionResponse": {
        "type": "object",
        "description": "API response type for a completed transaction.\n\nThis structure represents a completed transaction in the API response.\nIt contains all the relevant transaction details for display and tracking.\n\n# JSON Example\n\n```json\n{\n  \"id\": \"550e8400-e29b-41d4-a716-446655440000\",\n  \"pending_tx_id\": \"661e8400-e29b-41d4-a716-446655440001\",\n  \"account_id\": 1,\n  \"status\": \"broadcast\",\n  \"last_rejected_reason\": null,\n  \"kernel_excess_hex\": \"abc123...\",\n  \"sent_payref\": null,\n  \"sent_output_hash\": \"def456...\",\n  \"mined_height\": null,\n  \"mined_block_hash_hex\": null,\n  \"confirmation_height\": null,\n  \"broadcast_attempts\": 1,\n  \"created_at\": \"2024-01-15T10:30:00Z\",\n  \"updated_at\": \"2024-01-15T10:31:00Z\"\n}\n```",
        "required": [
          "id",
          "pending_tx_id",
          "account_id",
          "status",
          "kernel_excess_hex",
          "broadcast_attempts",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "account_id": {
            "type": "integer",
            "format": "int64",
            "description": "Account this transaction belongs to"
          },
          "broadcast_attempts": {
            "type": "integer",
            "format": "int32",
            "description": "Number of broadcast attempts made"
          },
          "confirmation_height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Block height where the transaction was confirmed (if confirmed)"
          },
          "created_at": {
            "type": "string",
            "description": "Timestamp when the transaction was created"
          },
          "id": {
            "type": "object",
            "properties": {
              "Tx_id": {
                "type": "integer"
              }
            }
          },
          "kernel_excess_hex": {
            "type": "string",
            "description": "Kernel excess in hexadecimal encoding"
          },
          "last_rejected_reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Reason for rejection if the transaction was rejected"
          },
          "mined_block_hash_hex": {
            "type": [
              "string",
              "null"
            ],
            "description": "Block hash where the transaction was mined (hex encoded, if mined)"
          },
          "mined_height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Block height where the transaction was mined (if mined)"
          },
          "pending_tx_id": {
            "type": "string",
            "description": "Reference to the original pending transaction"
          },
          "sent_output_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "Output hash for the sent transaction"
          },
          "sent_payref": {
            "type": [
              "string",
              "null"
            ],
            "description": "Payment reference if the transaction has been confirmed"
          },
          "status": {
            "type": "string",
            "description": "Current status of the transaction (completed, broadcast, mined_unconfirmed, mined_confirmed, rejected, canceled)"
          },
          "updated_at": {
            "type": "string",
            "description": "Timestamp when the transaction was last updated"
          }
        }
      },
...
      "DisplayedTransaction": {
        "type": "object",
        "description": "User-friendly transaction representation.",
        "required": [
          "id",
          "direction",
          "source",
          "status",
          "amount",
          "blockchain",
          "details",
          "lock_height"
        ],
        "properties": {
          "amount": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          },
          "blockchain": {
            "$ref": "#/components/schemas/BlockchainInfo"
          },
          "counterparty": {
            "type": "object",
            "properties": {
              "TariAddress": {
                "type": "string"
              }
            }
          },
          "details": {
            "$ref": "#/components/schemas/TransactionDetails"
          },
          "direction": {
            "$ref": "#/components/schemas/TransactionDirection"
          },
          "fee": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FeeInfo",
                "description": "Fee information (only populated for outgoing transactions)."
              }
            ]
          },
          "id": {
            "type": "object",
            "properties": {
              "Tx_id": {
                "type": "integer"
              }
            }
          },
          "lock_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "$ref": "#/components/schemas/TransactionSource"
          },
          "status": {
            "$ref": "#/components/schemas/TransactionDisplayStatus"
          }
        }
      },
...
      "FeeInfo": {
        "type": "object",
        "required": [
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "TransactionDirection": {
        "type": "string",
        "enum": [
          "incoming",
          "outgoing"
        ]
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Wallet API token, sent as 'X-API-Key: <token>'"
      },
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Wallet API token, sent as 'Authorization: Bearer <token>'"
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    },
    {
      "api_key": []
    }
  ],
...
}